Rename co_mha to co_re_mha
