+++ modules/__init__.py
@@ -1,1 +1,1 @@
-from .co_mha import CoReMultiheadAttention  # noqa: F401
+from .co_re_mha import CoReMultiheadAttention  # noqa: F401
